ResourceManager refactored towards a more generic solution. Fixed a bug in MaterialParser that prevented from loading the same material file twice.

+++ trunk/core-Renderer/Renderer.cpp
@@ -122,7 +122,8 @@ void Renderer::RenderingState::render(Renderer& renderer)
          (*it)->enable(true);
       }
 
-      activeCamera.setNearPlaneDimensions(renderer.m_viewportWidth, renderer.m_viewportHeight);
+      activeCamera.setNearPlaneDimensions((float)(renderer.m_viewportWidth), 
+                                          (float)(renderer.m_viewportHeight));
       // set the camera
       const D3DXMATRIX& viewMtx = activeCamera.getViewMtx();
       const D3DXMATRIX& camMtx = activeCamera.getGlobalMtx();
